Move `unused_qualifications` to the lints table. (#61)
We don't need to wait for MSRV support to use the lints table. The MSRV
`cargo check` commands won't fail, they'll just warn that there is an
unused manifest key but the CI will still pass. The stable toolchain
will do the actual linting.

This won't be a problem for projects higher up the stack either, as
Cargo doesn't complain about unused manifest keys of dependencies.