Search correct paths for fonts on Apple platforms. (#71)
Use the Foundation API to get the list of "Library" directories and look
in each of them for a "Fonts" directory.

This handles the need to search a variety of paths on macOS to find
fonts in the various standard locations like "~/Library/Fonts",
"/System/Library/Fonts", "/Library/Fonts", "/Network/Library/Fonts",
etc.

Fixes #50.