Tree style builder (#76)
- Builds on top of https://github.com/linebender/parley/pull/84 which
should be merged first to preserve history

## Changes made

- ### Entire unresolved style struct
     - Adds an "unresolved style struct" (`TextStyle`) to `style/mod.rs`
- The new `TreeBuilder` allows you to directly pass an entire style
struct (rather than just "changed styles" relative to a previous node in
the tree) as this is easier for integrating with Stylo which already
resolves inherited styles. So I have added an "unresolved" version of
the style struct to allow this API to still plug into the other parts of
Parley's style resolution functionality.
- Adds a corresponding`resolve_entire_style_set` method to
`LayoutContext` to convert `TextStyle` into `ResolvedStyle`.
- ### Moved code
- Moves the `RangedBuilder` from `context.rs` to a new `builder.rs`
module (as there are now two builders, justifying a separate module).
- Extract most of `RangedBuilder::build_into` into a standalone
`build_into_layout` function (`builder.rs`) that can be shared between
the ranged and tree builders.
- Moves the `RangedStyle` and `RangedProperty` types from
`resolve/range.rs` to `resolve/mod.rs`. These types are shared between
the `RangedBuilder` and the `TreeBuilder`.
- ### Tree builder
- Adds a `TreeBuilder` (also to `builder.rs`). This mostly delegates to
`TreeStyleBuilder`
- Add a `TreeStyleBuilder` (`resolve/tree.rs`). This is the vast
majority of the new code
- The `TreeStyleBuilder` implements HTML-style whitespace collapsing
(opt-in). This probably ought to become a style rather than being a flag
on the style builder.
- Updated swash example to optionally use the tree builder (depending on
command line opt)