CoWify style types (#128)
This:
- Converts parley style types to use `std::borrow::Cow` instead of
references
- Removes the lifetime parameter from `PlainEditor` and `PlainEditorOp`
by making them using `StyleProperty<'static>`

This isn't ideal from an API point of view, but does at least allow
`PlainEditor` to be stored within a struct.

It would be an option to leave the lifetime parameter in `PlainEditor`
and make uses who want to store it in a struct specify
`PlainEditor<'static>` themselves. I have not done so because I don't
think there's much use for the borrowed editor outside of toy examples.