Fixes #138 - placing inline boxes (#163)
When a inline box is placed at the beginning of the non-first line then
line breaking is wrong.

Lets have the following text:

```
Line1\nLine2
       |
       \- Inlinebox here
```

It produces the following items in layout:

[run, inlinebox, run]

But the line breaking occurs when the second run is processed. So
inlinebox ends on the previous line.
This PR fixes it by checking that current cluster_idx is mandatory break
and ends line before processing inline boxes