Fix Cursor::from_point to use the line's offset (#176)
This is effectively the opposite of #169.

I've also added a test. This test does use system fonts, and so is
written so as to be *relatively* independent of the font, assuming it
doesn't do any ligature fanciness.

In future, we'll likely want to us an in-repository small font (such as
roboto).