Initial version of snapshot testing (#173)
Hi,

This PR contains an initial version of snapshot testing.

By design it contains only a minimal API and usage inspired by vello
snapshot tests.
The ambition of this PR is not to do any test coverage, it contains only
two tests for demonstration purposes of the testing environment.

All the bells and whistles (html report, image diffs, not loading fonts
in every test) are now missing to make the code more straightforward.


# Usage

```bash
$ cargo test
```

If a test fails, you can compare images in `parley/tests/current`
(images created by the current test)
and `parley/tests/snapshots` (the accepted versions).

If you think that everything is ok, you can start tests as follows:

```bash
$ PARLEY_TEST="accept" cargo test
```

It will update snapshots of the failed tests.

# How it works

It reuses `tiny-skia-renderer` and renders the testing layout into an
image and compares it pixel by pixel.

This PR also adds fonts into the tests, to have tests independent on the
system fonts. I have chosen DejaVu fonts, but it was a random pick.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>