Cursor rework (#170)
This is a draft because some functionality (mostly around word/line
selection) is still missing. I also commented out the fix from #163
because the `skip_mandatory_break` flag is now gone and I haven't dug
into the logic from that PR.

Submitting this now so that people can take a look and see if it fixes
some of the known bugs, particularly around newline handling.

note: #166 should probably land first