Fix whitespace collapsing at the start of inline spans.
Generally, whitespace collapsing collapses sequences of whitespace into
a single space character. However, whitespace at the very start or end
of a span is elided entirely.

We use a flag to track if we are in this state. However, there was a bug
which was causing this flag to cleared when whitespace itself was pushed
Which caused start-of-span whitespace to appear as a single space if it
is pushed into the builder multiple chunks.

This patch prevents the flag from being cleared by whitespace to fix
that bug.