Ignore affinities in Selection::is_collapsed (#202)
Currently `is_collapsed` is false when a selection's anchor and focus
are at the same byte position, but have different affinities.

This changes `is_collapsed` to ignore the affinities, which appears to
be the sense of "collapsed" that is assumed throughout,
e.g.,`Selection::next_visual` currently requires doing the operation
twice when only affinities differ to actually move and
`PlainEditor::selected_text` currently returns `""` when only affinities
differ.