Handle inline boxes in cluster geometry and path
Inline boxes are currently silently ignored in some places, making text
selection wrong when they are used, as their advance is not used in
offset calculation and the run indices are counted wrong as the boxes
are skipped.

To see the issues currently present, insert the following in
`PlainEditor::update_layout` and run vello_editor.

```rust
builder.push_inline_box(crate::InlineBox {
    id: 0,
    index: 7,
    width: 50.0,
    height: 50.0,
});
```

This PR fixes the issues I've found, though the cursor position when the
cursor has a downstream affinity right on an inline box is still wrong,
I believe. For example, in "AB" with an inline box between "A" and "B",
when the cursor is between the two characters with downstream affinity,
still shows up on the right edge of the A.