Don't require `Eq` for `FontSettings` (#206)
The code currently requires that `T` be `Eq` for `FontSettings<T>`.

This is okay when using `swash::Setting` as that unconditionally defines
`Eq` for `Setting<T>` when `T: PartialEq` and doesn't check for `Eq`.

When using `skrifa::setting::Setting`, then `Eq` is only defined for
`Setting<T>` when `T: Eq`.

We use `Setting` (and `FontSettings`) for `FontVariation` which is a
`Setting<f32>` and therefore is not `Eq` due to the `f32`.