fontique: Move creation of attributes from fontconfig (#212)
This exposes `from_fontconfig` functions on `FontStretch`, `FontStyle`,
and `FontWeight` rather than only defining them locally within the
fontconfig code.

This is in preparation for a change similar to #209 where these types
and others get pulled into a shared vocabulary crate and this sort of
constructor helper would be useful to others.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>