Add `PlainEditor` method to get IME cursor area (#224)
The area reported is the area of text the user is currently editing. It
is usually used by platforms to place a candidate box for IME near it,
while ensuring it is not obscured.

The implementation here ensures the area is usually near the focus, that
the area contains at least some surrounding context (in case the
platform places something to the side of the area, instead of above or
under), and that the IME candidate box usually does not need to jump
around when the IME starts or continues composing.