Fix linebreaking for lines without text runs (#249)
It turns out that there were actually two bugs that caused the behavior
as described in #246.

The first bug happens when a line ends with a text run. In this
scenario, the text run is repeated at the start of the next line, so
that any remaining clusters can be rendered. However, when there are no
more remaining clusters to render, the text run essentially uses
clusters from the next run. In the example image in #246, this causes
the 'B' to be added to the second line.

The second bug happens when a line that does not contain any text runs
is committed. In this case, the cluster range in the current line state
is bumped, even though no clusters were actually processed. This is the
reason that the 'B' wasn't added to the third line in the image in #246.

I've added a test that tests a scenario with a full width inline box
both when the previous line ends with an inline box and when it ends
with a text run.

I'm not fully satisfied with the solution in this PR, because it feels
rather heavy-handed. It is already unclear to me when both existing
conditions in the invalid cluster check even apply, so just adding
another one makes it even harder to understand.

Fixes #246