Fix RTL text alignment; make start/end alignment depend on text direction (#250)
Also adds `Alignment::{Left, Right}` for bidi-independent alignment.

Various issues around trailing white space handling are fixed. The main
issue is that in the RTL case, trailing white space is on the left,
whereas in the LTR case it's on the right. As increasing the
x-coordinate is to the right, the white space doesn't automatically hang
in the RTL case.

Other issues are related to bidi endianness as well: e.g., trailing
white space is either the first or the last cluster in visual order in
the RTL and LTR cases respectively.

This doesn't test mixed-bidi text yet.