Calculate minimum and maximum content width (#259)
This adds `min_content_width` and `max_content_width` functions to the
layout. These functions provide an lower and upper bound on the layout
width. The content widths are trailing white space aware, which means
that they correspond to `layout.width` and not `layout.full_width`.

Follow-up work includes using the `max_content_width` to ensure the
`max_advance` is always capped. This will become useful for floated
boxes (#99), because those require Parley to align individual lines
based on their `max_advance`, which in turn requires a reasonable upper
bound on `max_advance`.