Fix and simplify `unjustify`, remove `LineData::alignment` (#271)
`unjustify` doesn't mirror `align`, so it performs the wrong
calculations, and re-line-breaking or re-aligning currently lead to the
wrong result after a justified alignment.

As the alignment function requires further iteration (e.g. for floated
boxes), this patch removes the specialized `unjustify` implementation
and instead uses the same code for aligning and unjustifying. This is a
trade-off between efficiency and correctness. If this is merged in, we
can always split the two again in the future once alignment has
stabilized.

This stores the width the layout was aligned to, which is necessary for
unjustifying.