Account for inline boxes when collapsing after newlines (#280)
Followup (small bug fix) to
https://github.com/linebender/parley/pull/254 which didn't take inline
boxes into account. That PR was checking whether "the last character is
whitespace" without checking whether the last thing pushed into the
layout was text.

So if the sequence of items pushed to a layout was something like:

- inline box
- whitespace
- inline box
- whitespace
- inline box
- whitespace
- inline box

(an example which would be "status badges" in Github readme's which are
typically images separated by whitespace characters)

Then all but the first "whitespace" would get collapsed as it would be
incorrectly detected as directly adjacent to the preceding whitespace
(even though it shouldn't because there is an inline box in between).

This PR fixes this issue by tracking the kind of the last item pushed.

---------

Signed-off-by: Nico Burns <nico@nicoburns.com>