Display selected newlines as whitespace in the selection highlight (#296)
This is the behavior that I see in Chrome and in GTK and QT apps (at
least on Linux). If you select the newline at the end of a line, it's
displayed as some trailing whitespace:

![image](https://github.com/user-attachments/assets/65377e35-d680-4c43-80ef-a49bbd5aaecc)

![image](https://github.com/user-attachments/assets/f81a772a-4e97-4d5b-bbc1-fc0a3a659db5)

This implements that behavior in Parley. This is an improvement over the
existing behavior, which forces all selection boxes to be at least 8px
wide, even if the text being selected is narrower than that (e.g. you
select a single letter "i" at a smallish font size).

As stated in the TODO comment, ideally this whitespace would be as wide
as the space (U+0020) character in the run's font, but we currently have
no way to pass that information down.