Add APIs for overriding fonts' metadata at registration time and unregistering fonts (#312)
Resolves https://github.com/linebender/parley/issues/117, implementing
the APIs requested there.

The first API added is `Collection::clear`, which resets the list of
registered fonts and associated data (custom generic families and
fallbacks, etc).

The second API addition is a helper struct which can be passed as a
parameter to `Collection::register_fonts`, which allows the caller to
override the font's family name and other metadata. I originally thought
about passing a callback which could be called once per font in the
provided data, but I think that's overkill and makes the most common use
case a lot more cumbersome.

This is necessary for the egui work, since egui lets users specify the
names of registered fonts.