Pass the line index to the `Selection`'s geometry methods (#318)
In order to render text with a background in a way that matches CSS, one
needs to draw it a line at a time, background then selection then
foreground:


![image](https://github.com/user-attachments/assets/d00e9ee5-25b2-4222-af19-25cf424002e8)

If you don't know which line a selection rectangle belongs to, you can't
get the effect where each line is layered atop the one before it, since
you can only draw the entire selection at once, and hence the entire
background at once.

I've also added a `selection_geometry_with` function to
`PlainEditor`--while updating the rest of the codebase, I noticed it was
missing and could be useful (e.g. in the `vello_editor` example.