Add `raw_text` function to `PlainEditor` (#317)
I'm not sure it's always a good idea to exclude the IME preedit region
from the externally exposed `PlainEditor` text buffer. In any case,
having access to the raw buffer will make it easier to implement
`InputConnection` on Android. @DJMcNab if you'd rather wait and review
that code when it's ready, to see if it suggests other changes that
should be made to the `PlainEditor` API, that's fine with me; I'll go
ahead and use this branch in the android-view demo/test app.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>