Add `is_done` method to `BreakLines` (#319)
With the text wrapping styles implemented, I can implement ellipsis
truncation in egui by simply creating another layout for the ellipsis.
However, to know whether to show the ellipsis, I need to know if there's
any text left to wrap. It would be nice to be able to do that without
having to actually break another line and then revert it.