Don't reset cursor affinity on newlines (#331)
Resolves https://github.com/linebender/parley/issues/298.

This bit of code was apparently here to ensure the cursor moved down a
line when you typed a newline, but I've tested it (with both cursor
affinities, and around wrapped lines) and didn't observe any change in
behavior with it removed.

Since having this code here breaks RTL text navigation, and doesn't seem
to do anything else, I think it can be removed now.

~~There is some weird (probably incorrect) behavior around inserting
newlines within an RTL span--as far as I can tell, all of that behavior
is pre-existing, and occurs with or without this code present. I'll file
a new issue about that.~~ (After testing how browsers behave in this
scenario, this appears to just be how it works. Bidirectional text is
weird.)