fontique: Fix panic on macOS in debug mode. (#335)
This commit fixes a panic on macOS by enabling the "relax-sign-encoding"
feature for the "objc2" crate.

Relevant issues in the objc2 crate:
https://github.com/madsmtm/objc2/issues/566
https://github.com/madsmtm/objc2/pull/567

I made sure to use version of objc2 that the current objc2-* variations
use, so that we don't duplicate the crate.
![Screenshot 2025-04-11 at 8 51
30 PM](https://github.com/user-attachments/assets/590301c3-4b08-47e6-8ac6-f81f39d8fea6)

---
Additional Info:


System Info: Mac Mini M4, 16 GB of RAM.

Before the fix, when running the vello-editor or any example:
![Screenshot 2025-04-11 at 9 12
02 PM](https://github.com/user-attachments/assets/45a98667-25a8-4da4-98e5-b5d414c0b9e3)

After the fix:
![Screenshot 2025-04-11 at 9 14
29 PM](https://github.com/user-attachments/assets/4864d4e1-5c54-47b2-bcd8-d0788fa76d07)