Add option to quantize vertical layout metrics. (#344)
## Background

Vertical metrics are no longer rounded since
https://github.com/linebender/parley/pull/297 and so the rounding needs
to happen in the renderer. The upside is that there's potential for more
control, the downside is that it's another footgun. I initially created
documentation and updated examples in #343 but after going through that
I came to the conclusion that asking users to do that will be a major
inconvenience, and likely will lead to incorrect results.

## Solution

This PR here takes a different approach. We bring back quantization but
make it optional. That way advanced users who know what they're doing
can still get access to the raw fractions. However, for most users we
can recommend just using our quantization which will automatically take
care of the blurry text issue. The examples will also use the much
simpler quantization option.

## Implementation details

Now as pointed out in #297 our old quantization strategy was flawed
where rounding errors accumulated and weren't dealt with. I didn't want
to just bring that back so instead I played around with Chrome a bunch
to see how it does line height handling. This PR now contains a solution
that mimics Chrome.

We now only round `baseline`, `min_coord`, and `max_coord` in the
exported metrics. There's more rounding internally, but only temporarily
to calculate the aforementioned. So e.g. exported `ascent` will still be
the raw fractional. For the line's `y` position, when the accumulated
overflow reaches over half a pixel we will either overflow to the
previous line or start with a gap, depending on the direction of the
accumulation.

## Examples

As seen in the example below (*16.0px font size, 1.3em line height
(20.8px)*), we now match Chrome's behavior of overflowing the **1st,
4th, 9th, and 14th** line over the previous line. Best spotted by
looking at the edges of the red/green selection boxes. Also note that
despite the perceived line height fluctuating between lines due
positional changes, the rendered line height does not change. Easy to
see via the selection boxes, which are always the same height. (*They're
colored spans in Chrome, but colored spans are the same height as
selection boxes until line height is greater than ascent + descent,
which is not the case in this example - making it a valid comparison.*)

![16 0px-1
3em-2fps](https://github.com/user-attachments/assets/d80e3a2a-29b5-4166-933a-0172dbad3029)

Here's another example (*16.0px font size, 1.33125em line height
(21.3px)*) showing a scenario where gaps are formed due to the line
height rounding going in the other direction. Also the line height is
greater than ascent + descent, so the gaps are actually visible with
selection boxes.

| Chrome | Swash |
| --- | --- |
| ![16 0px-21
3px-chrome](https://github.com/user-attachments/assets/ea0fb537-1806-4fa2-b993-4f566b7260d6)
| ![16 0px-21
3px-swash](https://github.com/user-attachments/assets/32aadfde-670d-4a4b-a81b-15d3fd5a7c07)
|

Here's a bunch more examples:

| Notes | Chrome | Swash |
| --- | --- | --- |
| Mixing of fonts and sizes, 1.99em line height |
![mixed-fonts-chrome](https://github.com/user-attachments/assets/9d319bb7-7797-4e08-9541-c2976a1e9c30)
|
![mixed-fonts-swash](https://github.com/user-attachments/assets/a28c446a-9827-4a84-ba8f-b45b6f5e4173)
|
| 16px 0.83em (13.28px) | ![16 0px-0
83em-chrome](https://github.com/user-attachments/assets/e7314452-db2a-4980-adb7-8140820fa301)
| ![16 0px-0
83em-swash](https://github.com/user-attachments/assets/0d6a34ac-c1d9-469a-9f16-82032088328e)
|
| 19px 0.83em (13.28px) | ![19 0px-0
83em-chrome](https://github.com/user-attachments/assets/f4950b1c-7b65-4276-8acb-0d1e9db386c4)
| ![19 0px-0
83em-swash](https://github.com/user-attachments/assets/61315181-47b7-4122-b5a3-6c824f38c29e)
|

## Mismatches

However, it's not always a perfect match. Fractional font sizes are
rendered at a different size, which probably causes the layout changes
seen in the 18.5px example. Also, when going to very small line heights,
starting from around 0.65em Chrome has some additional logic to keep
perceived line height higher. I'm not exactly sure what it is, but it
accumulates by the time you get to tiny line heights like 0.25em. It's
not related to the rounding of the line height, because in the 0.25em
case the line height is exactly 4px. Anyway, not super important right
now, but might be worth figuring out later.

| Notes | Chrome | Swash |
| --- | --- | --- |
| 18.5px 0.83em (13.28px) | ![18 5px-0
83em-chrome](https://github.com/user-attachments/assets/11980f67-18c7-419b-8006-809712506b61)
| ![18 5px-0
83em-swash](https://github.com/user-attachments/assets/9d3b44f0-bfd6-4ff5-86b1-08c304a468c0)
|
| 16px 0.65em (10.4px) | ![16 0px-0
65em-chrome](https://github.com/user-attachments/assets/c5aa3452-0f1e-4eb7-9326-167dade930e3)
| ![16 0px-0
65em-swash](https://github.com/user-attachments/assets/9acff7c9-f933-42b9-8521-3f43381e871c)
|
| 16px 0.25em (4px) | ![16 0px-0
25em-chrome](https://github.com/user-attachments/assets/3016809f-ce1f-4b6c-b089-e4c20364797e)
| ![16 0px-0
25em-swash](https://github.com/user-attachments/assets/cf3a92d9-727c-4bc0-8601-7977fa94a33f)
|

---

Alternative to and closes #343.