Don't cache min/max content widths in `Layout` (#353)
Caching the min/max content widths of a `Layout` was previously done
using a `OnceCell`. This makes the entire `Layout` struct neither `Send`
nor `Sync`, and means we have to wrap it in a `Mutex` if we want those
properties.

Because of the requirement to use a `Mutex`, API consumers that never
use the min/max content widths (or could easily cache them themselves)
but *do* need `Layout` to be `Send` or `Sync` end up paying for a
feature that they never use.

There are a few different solutions here:
- Use an `Option` instead of a `OnceCell` and have the content size
methods take `&mut self`
- Cache the sizes in an `AtomicU64` or `OnceLock` instead
- Don't cache the sizes at all and leave it up to the API consumer

Based on [discussion on
Zulip](https://xi.zulipchat.com/#narrow/channel/205635-parley/topic/Should.20the.20.60Layout.60.20type.20return.20to.20being.20.60Sync.60.3F/near/508638378),
the last option seems to be the preferred one, so that's what I've gone
with here. If the caching is going to be a leaky abstraction anyway,
it's probably best not to bother.

I've also added a test to ensure that `Layout` always implements `Send +
Sync` going forward.