Test that all the builders produce the same results. (#366)
This will help ensure that the builders won't diverge in behavior as we
continue modifying them. It also tests the reusability of
`LayoutContext`.

There are known cases where the builders do currently diverge. For
example:
* Using only default styles.
Would be good to have it catch the builders disagreeing on default font
size or missing scale multipliers. While adding this test is trivial
within the framework of this PR, unfortunately the solution isn't super
trivial. So I will add the default style test along with the code fix in
a separate PR.
* When adding a span with the same style as the currently active span.
  For example see #365.
However, this gets into questions about supporting things like
margin/padding on spans and how exactly the common core span API should
look like.

I wanted to keep this PR here limited to the cases that don't require
any builder modification. So that we can at least start protecting the
partial commonality that the builders have. This will also mean that the
follow-up PRs that do change builder code won't be burdened by the noise
of adding a bunch of this testing framework code.