Use the correct family ID when adding font aliases (#380)
Discovered this as part of #378; splitting it off because it affects all
the fontique backends.

When adding font aliases, the code was not actually using the family ID
we passed in, but creating a fresh family ID each time. This meant that
font aliases were all mapped to nonexistent family IDs, and all queries
for a font by one of its aliases would fail.

I'm a bit concerned that this has the potential to cause regressions or
other general weirdness, since all font aliases went unused due to this
bug. Not sure how to test that other than to just give it a spin.