Add unregister font method (#395)
***I am new to Parley, so I am not sure if this functionality already
exists! 🙏 ***



From [#parley > Supporting font unloading @
💬](https://xi.zulipchat.com/#narrow/channel/205635-parley/topic/Supporting.20font.20unloading/near/531473904):

> I have a use case for font unloading/replacing. At a high level, when
the page loads, we eagerly install a given font subset. Lazily, we load
the full font to replace the subset at a later time. Separately, I would
love to unload a font that is no longer needed in memory.
> 
> IIUC, Fontique's collection doesn't support fine grained removal
besides
[Collection::clear](https://github.com/linebender/parley/blob/b9852143e6b78f10f439bfc5a2f9e26fa9ae6f12/fontique/src/collection/mod.rs#L189-L194).
I think this is a little cumbersome for font unloading because it means
I need to replicate the collection's state in consuming code (e.g.
consumer code "remembers" collection state, when we need to unload, we
clear the collection, and add back the remembered state except the font
we want to unload).
> 
> I also considered using
[register_font](https://github.com/linebender/parley/blob/b9852143e6b78f10f439bfc5a2f9e26fa9ae6f12/fontique/src/collection/mod.rs#L175-L187)
to "overwrite" any existing face in the family of the same style.
Currently, I think adding the same face twice to a collection results in
the collection increasing to store the second face, but without it
actually being used.